security

+++ src/main/java/com/techpal/sn/repository/RoleRepository.java
@@ -1,14 +1,13 @@
 package com.techpal.sn.repository;
 
-import java.util.Optional;
-
+import com.techpal.sn.models.ERole;
+import com.techpal.sn.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.techpal.sn.models.ERole;
-import com.techpal.sn.models.Role;
-
+import java.util.Optional;
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-	Optional<Role> findByName(ERole name);
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+      Optional<Role> findByName(ERole name);
+
 }
